RePlug/requirements.txt
# requirements.txt

# General libraries
numpy
faiss-cpu
openai
torch
transformers
boto3
botocore
datasets
deepspeed


# Additional libraries
requests

# You might need to specify specific versions if necessary for compatibility, for example:
# torch==1.10.0
# transformers==4.15.0
# faiss-cpu==1.7.0
